+++ NEWS.md
@@ -1,4 +1,4 @@
-# etl 0.3.2
+# etl 0.3.3
 
 * Moved `is.etl` to main documentation for `etl` (30dee378)
 * Fixed typo in DESCRIPTION (4e77fba2)